+++ cmem_plugin_base/dataintegration/plugins.py
@@ -19,6 +19,7 @@ class WorkflowPlugin:
             or none (no outgoing connection).
         """
 
+
 class TransformPlugin:
     """
     Base class of all transform operator plugins.
@@ -27,8 +28,8 @@ class TransformPlugin:
     def transform(self, inputs: Sequence[Sequence[str]]) -> Sequence[str]:
         """
         Transforms a collection of values.
-        :param inputs: A sequence which contains as many elements as there are input operators for this transformation.
-                       For each input operator it contains a sequence of values.
+        :param inputs: A sequence which contains as many elements as there are input
+            operators for this transformation.
+            For each input operator it contains a sequence of values.
         :return: The transformed values.
         """
-        pass